# core/generated/swagger-ui/schema-registry-api-spec.yaml
...
swagger: "2.0"
info:
  version: "v1"
  title: "Confluent Schema Registry"
schemes:
- "http"
- "https"
paths:
...
  /schemas/ids/{id}:
    get:
      summary: "Get the schema string identified by the input ID."
      description: ""
      operationId: "getSchema"
      consumes:
      - "application/vnd.schemaregistry.v1+json"
      - "application/vnd.schemaregistry+json"
      - "application/json"
      - "application/octet-stream"
      produces:
      - "application/vnd.schemaregistry.v1+json"
      - "application/vnd.schemaregistry+json; qs=0.9"
      - "application/json; qs=0.5"
      parameters:
      - name: "id"
        in: "path"
        description: "Globally unique identifier of the schema"
        required: true
        type: "integer"
        format: "int32"
      - name: "format"
        in: "query"
        required: false
        type: "string"
      - name: "fetchMaxId"
        in: "query"
        required: false
        type: "boolean"
        default: false
      - name: "Authorization"
        in: "header"
        required: false
        type: "string"
      - name: "Cookie"
        in: "header"
        required: false
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/SchemaString"
        404:
          description: "Error code 40403 -- Schema not found\n"
        500:
          description: "Error code 50001 -- Error in the backend data store\n"
  /schemas/ids/{id}/subjects:
    get:
      summary: "Get all the subjects associated with the input ID."
      description: ""
      operationId: "getSubjects"
      consumes:
      - "application/vnd.schemaregistry.v1+json"
      - "application/vnd.schemaregistry+json"
      - "application/json"
      - "application/octet-stream"
      produces:
      - "application/vnd.schemaregistry.v1+json"
      - "application/vnd.schemaregistry+json; qs=0.9"
      - "application/json; qs=0.5"
      parameters:
      - name: "id"
        in: "path"
        description: "Globally unique identifier of the schema"
        required: true
        type: "integer"
        format: "int32"
      - name: "Authorization"
        in: "header"
        required: false
        type: "string"
      - name: "Cookie"
        in: "header"
        required: false
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              type: "string"
            uniqueItems: true
        404:
          description: "Error code 40403 -- Schema not found\n"
        500:
          description: "Error code 50001 -- Error in the backend data store\n"
  /schemas/ids/{id}/versions:
    get:
      summary: "Get all the subject-version pairs associated with the input ID."
      description: ""
      operationId: "getVersions"
      consumes:
      - "application/vnd.schemaregistry.v1+json"
      - "application/vnd.schemaregistry+json"
      - "application/json"
      - "application/octet-stream"
      produces:
      - "application/vnd.schemaregistry.v1+json"
      - "application/vnd.schemaregistry+json; qs=0.9"
      - "application/json; qs=0.5"
      parameters:
      - name: "id"
        in: "path"
        description: "Globally unique identifier of the schema"
        required: true
        type: "integer"
        format: "int32"
      - name: "Authorization"
        in: "header"
        required: false
        type: "string"
      - name: "Cookie"
        in: "header"
        required: false
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/SubjectVersion"
        404:
          description: "Error code 40403 -- Schema not found\n"
        500:
          description: "Error code 50001 -- Error in the backend data store\n"
  /schemas/types:
    get:
      summary: "Get the schema types supported by this registry."
      description: ""
      operationId: "getSchemaTypes"
      consumes:
      - "application/vnd.schemaregistry.v1+json"
      - "application/vnd.schemaregistry+json"
      - "application/json"
      - "application/octet-stream"
      produces:
      - "application/vnd.schemaregistry.v1+json"
      - "application/vnd.schemaregistry+json; qs=0.9"
      - "application/json; qs=0.5"
      parameters:
      - name: "Authorization"
        in: "header"
        required: false
        type: "string"
      - name: "Cookie"
        in: "header"
        required: false
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              type: "string"
            uniqueItems: true
        500:
          description: "Error code 50001 -- Error in the backend data store\n"
...
definitions:
...
  SchemaReference:
    type: "object"
    properties:
      name:
        type: "string"
      subject:
        type: "string"
      version:
        type: "integer"
        format: "int32"
  SchemaString:
    type: "object"
    properties:
      schemaType:
        type: "string"
        description: "Schema type"
      schema:
        type: "string"
        description: "Schema string identified by the ID"
      references:
        type: "array"
        description: "Schema references"
        items:
          $ref: "#/definitions/SchemaReference"
      maxId:
        type: "integer"
        format: "int32"
        description: "Maximum ID"
...
  SubjectVersion:
    type: "object"
    properties:
      subject:
        type: "string"
      version:
        type: "integer"
        format: "int32"